Update traefik.toml
This configuration defines entrypoints for both HTTP and HTTPS, as well as a web interface for the status page. It also sets up the Docker provider to watch the Docker socket and automatically discover services and containers.

Next, the configuration defines routers for each service, using the frontends section. Each router is associated with a specific hostname and uses a specific backend to handle incoming requests.

Finally, the configuration defines the backends for each service using the backends section. Each backend is associated with a specific service and contains the URL of the service that will handle incoming requests.

This configuration sets up Traefik to route incoming requests to the correct service based on the hostname in the request. It also allows for flexible scaling of services as your needs change, by allowing you to add or remove servers as necessary.